requirements.txt
fastapi
uvicorn[standard]
python-dotenv
gunicorn
psycopg2-binary
sqlalchemy
